code: fixed event size type

+++ code/header/general.h
@@ -45,7 +45,7 @@ LIBRG_API librg_event_type  librg_event_type_get(librg_world *world, librg_event
 LIBRG_API int64_t           librg_event_owner_get(librg_world *world, librg_event *event);
 LIBRG_API int64_t           librg_event_entity_get(librg_world *world, librg_event *event);
 LIBRG_API char *            librg_event_buffer_get(librg_world *world, librg_event *event);
-LIBRG_API size_t            librg_event_size_get(librg_world *world, librg_event *event);
+LIBRG_API int32_t           librg_event_size_get(librg_world *world, librg_event *event);
 LIBRG_API void *            librg_event_userdata_get(librg_world *world, librg_event *event);
 
 // =======================================================================//
